Relocate gravity client
